% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archs4utils.R
\name{getVersions}
\alias{getVersions}
\title{List all versions of ARCHS4.}
\usage{
getVersions()
}
\value{
information of ARCHS4 versions.
}
\description{
List all versions of ARCHS4.
}
\examples{
getVersions()
}
