% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Say hello to the world.}
\usage{
hello(x, y)
}
\arguments{
\item{x}{Dummy variable 1.}

\item{y}{Dummy variable 2.}
}
\value{
The string it printed out.
}
\description{
Say hello to the world.
}
\examples{
hello(1, 1)
hello(10, 1)
}
